requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SpeechRecognition==3.10.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.0